Hot fix none error and adding the edit_task function
Fix the "none" error that was appearing in the function delate_task(). This happened because the print was called twice, this means that in the show_tasks() function it already printed the tasks and I sent the function to print:
print(show_tasks())
That's why there was the error appeared.

I also created the edit_tasks() function which is similar to the delete_tasks() function, but this function allows you to edit already existing tasks.